Refactor shared flow trigger logic into a composable
