+++ packages/flow-trigger-bundle/package.json
@@ -32,7 +32,7 @@
 			{
 				"type": "interface",
 				"name": "system-flow-trigger-form",
-				"source": "src/system-flow-trigger-form/index.ts"
+				"source": "src/_system/system-flow-trigger-form/index.ts"
 			}
 		],
 		"host": "^10.10.0"
@@ -49,7 +49,7 @@
 		"vue": "^3.4.38"
 	},
 	"dependencies": {
-		"@directus/format-title": "10.1.2",
+		"@directus/format-title": "11.0.0",
 		"@directus/types": "11.1.1",
 		"vue-i18n": "9.13.1"
 	}
