+++ packages/flow-trigger-bundle/tsconfig.json
@@ -2,7 +2,8 @@
 	"compilerOptions": {
 		"target": "ES2019",
 		"lib": ["ES2019", "DOM"],
-		"moduleResolution": "node",
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext",
 		"strict": true,
 		"noFallthroughCasesInSwitch": true,
 		"esModuleInterop": true,
